pyproject.toml
[project]
name = "comarketer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mlflow[databricks]>=3.1.3",
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
    "databricks-langchain>=0.5.0",
    "databricks-sdk>=0.50.0",
    "flask>=3.0.0",
    "asgiref>=3.7.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "databricks-langchain[memory]",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.8.0"]
